Move ConvertM into own module, make convertPrim failable
